refactor: move VersionedConstants macro, errors from Blockifier to StarknetAPI
