+++ crates/blockifier/src/versioned_constants_test.rs
@@ -1,5 +1,9 @@
+use std::path::PathBuf;
+
 use glob::{glob, Paths};
 use pretty_assertions::assert_eq;
+use starknet_api::block::StarknetVersion;
+use starknet_infra_utils::compile_time_cargo_manifest_dir;
 
 use super::*;
