+++ crates/starknet_api/src/lib.rs
@@ -23,6 +23,7 @@ pub mod test_utils;
 pub mod transaction;
 pub mod transaction_hash;
 pub mod type_utils;
+pub mod versioned_constants_logic;
 
 use std::num::ParseIntError;
